requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.1.8
Django==5.1.6
django-debug-toolbar==5.0.1
django-filter==25.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
isort==6.0.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
